backend/requirements.txt
Django
djangorestframework
paho-mqtt
django-cors-headers
Pillow
numpy
